Renamed and updated component
